refactor(dsp): remove catalog spi
