+++ settings.gradle.kts
@@ -65,7 +65,6 @@ include(":core:data-plane-selector:data-plane-selector-core")
 include(":data-protocols:dsp:dsp-api-configuration")
 include(":data-protocols:dsp:dsp-catalog:dsp-catalog-api")
 include(":data-protocols:dsp:dsp-catalog:dsp-catalog-http-dispatcher")
-include(":data-protocols:dsp:dsp-catalog:dsp-catalog-spi")
 include(":data-protocols:dsp:dsp-catalog:dsp-catalog-transform")
 include(":data-protocols:dsp:dsp-transfer-process:dsp-transfer-process-transform")
 include(":data-protocols:dsp:dsp-negotiation:dsp-negotiation-api")
